Move CFP in app folder, use appTemplate
